backend/pyproject.toml
[project]
name = "ravedar-backend"
version = "0.1.0"
requires-python = ">=3.14,<3.15"
dependencies = [
  "fastapi[standard]==0.135.2",
  "PyYAML==6.0.3",
  "psycopg[binary]==3.3.3",
  "sqlalchemy==2.0.48",
  "uvicorn==0.42.0",
]

[tool.uv]
package = false
